requirements.txt
yt-dlp>=2024.1.0
google-genai>=0.3.0
google-cloud-texttospeech>=2.14.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib20.0
youtube-transcript-api>=0.6.0
gTTS>=2.4.0
moviepy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
PyYAML>=6.0.0

